feat: add compose NavHost
Introduce a compose navigation graph to the app. Uses a sealed class to model the app screens separate to the graph itself following the example in the compose documentation. The launch list has been set as the default destination.

The detekt configuration for the function naming rule has been updated to ignore all functions with marked with @Composable. As this is more reliable and maintainable than updating the list of excluded files.