refactor: update project navigation to use functions as described in the docs
Change how the jetpack compose navigation is defined to follow the 'Type safety in Kotlin DSL and Navigation Compose' guide. The key difference being the use of extension functions on the compose navigation to create an expressive DSL for declaring the navigation graph.

These functions and removing the 'Screens' sealed class allows the information about the routes and arguments to be encapsulated into a single file. The detail view model makes use of a wrapper class to prevent navigation arguments keys from being used in the view model.

https://developer.android.com/guide/navigation/navigation-type-safety