test: add launch list test
Add a test to check a rocket launch name is displayed to the user. This required replacing production components with fakes for testing using Hilt. Such as using test dispatchers, an in memory database and fake API client.

Packaging options have been added to the app module as using `ilt-android-compiler` in test and `hilt-compiler` when running the app caused a duplicate file to be created.